# Optimization Attempt 2: Word Registers