Before:
```
Measuring the speed of 500 boot times.
Median boot time: 43.6ms.
Maximum boot time: 46.8ms.
Minimum boot time: 43.0ms.
```

After:
```
Measuring the speed of 500 boot times.
Median boot time: 48.7ms.
Maximum boot time: 53.7ms.
Minimum boot time: 48.2ms.
```